requirements.txt
opencv-python
Pillow
torchmetrics
easydict